+++ orchestration/src/test/java/com/sap/ai/sdk/orchestration/LLMModuleResultDeserializerTest.java
@@ -2,7 +2,7 @@ package com.sap.ai.sdk.orchestration;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import com.sap.ai.sdk.orchestration.model.LLMModuleResultSynchronous;
+import com.sap.ai.sdk.orchestration.model.LLMModuleResult;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 
@@ -41,7 +41,7 @@ class LLMModuleResultDeserializerTest {
         """;
 
     var json = String.format(JSON, choices);
-    var result = OrchestrationClient.JACKSON.readValue(json, LLMModuleResultSynchronous.class);
-    assertThat(result).isExactlyInstanceOf(LLMModuleResultSynchronous.class);
+    var result = OrchestrationClient.JACKSON.readValue(json, LLMModuleResult.class);
+    assertThat(result).isExactlyInstanceOf(LLMModuleResult.class);
   }
 }